优化更精准的状态

+++ util/system/bot.go
@@ -14,7 +14,7 @@ func Bot(content string) {
 	botFromUrl(content, SettingData.BotUrl)
 }
 
-func BotErr(code int, errString string) {
+func BotErr(code int64, errString string) {
 	botFromUrl(fmt.Sprintf("(code %d)%s", code, errString), SettingData.BotUrl)
 }
